requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
cycler==0.11.0
easyocr==1.4.1
Flask==2.1.0
fonttools==4.31.2
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pdf2image==1.16.0
Pillow==8.2.0
pyparsing==3.0.7
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tifffile==2022.3.25
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
